pyproject.toml
[project]
name = "atrax"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "browserdriver>=0.0.6",
    "django>=5.2",
    "django-ninja>=1.4.1",
    "requests>=2.32.3",
]
